requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
numpy==2.3.2
python-dotenv==1.1.1
pytube==15.0.0
requests==2.32.4
urllib3==2.5.0
